Move SearchResults classes to their own file.
